Thea-engine/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
chromadb==0.4.18
sentence-transformers==2.2.2
PyPDF2==3.0.1
pdf2image==1.16.3
pytesseract==0.3.10
easyocr==1.7.0
Pillow==10.1.0
numpy<2.0
python-dotenv==1.0.0
openai>=1.0.0
opencv-python>=4.8.0
PyMuPDF>=1.23.0
python-docx==1.1.0
openpyxl==3.1.2
xlrd==2.0.1
python-pptx==0.6.23
